+++ libraries/AD8495/CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.1.1] - 2024-11-07
+- add **float getTemperatureF()**
+- add default offset of 0.
+- add examples, (offset, plotter, Fahrenheit)
+- add **float getSetPointVoltage(float Temperature)** helper function
+- some optimizations
+- update readme.md
+- fix typo 9496 => 8496
+
+
 ## [0.1.0] - 2024-11-07
 - initial version